bpf: Do not bypass conntrack if running kube-proxy+vxlan+endpoint routes
This commit fixes a broken packet path on pod --> remote pod via VIP
when running with encapsulation and per-endpoint routes, with kube-proxy
handling DNATing. The following details the broken packet path:

- SYN leaves lxc device for stack             with pod1IP -> VIP
- SYN goes through cilium_host and cilium_net with pod1IP -> pod2IP
- SYN is sent to second node through tunnel   with pod1IP -> pod2IP
- SYN+ACK arrives from tunnel on cilium_vxlan with pod2IP -> pod1IP
- SYN+ACK is processed by ipvX_local_delivery
  and redirected in BPF directly to the lxc
  device
- SYN+ACK arrives on lxc device               with pod2IP -> pod1IP
- RST is sent because pod2IP doesn't match VIP

Because we use redirect() in BPF, the SYN+ACK packet bypasses most of
the stack (including conntrack) on its way from cilium_vxlan to the lxc
device. By bypassing conntrack, we also bypass the reverse DNAT.

The bug doesn't occur when per-endpoint routes are disabled because, in
that case, the SYN is SNATed to the cilium_host IP before leaving
through the tunnel. The SYN+ACK packet therefore takes a different path
and goes through conntrack.

Fixes: https://github.com/cilium/cilium/issues/14646
Co-authored-by: John Fastabend <john.fastabend@gmail.com>
Signed-off-by: Paul Chaignon <paul@cilium.io>